Add several functions, update header and printf array, rot13 still se…

+++ print_p.c
@@ -5,36 +5,36 @@
  * @ptr_val: address
  * Return: count
  */
-int print_px(int count, unsigned long long int ptr_val)
+int print_px(int count, unsigned long int ptr_val)
 {
-	unsigned long long int temp_num = ptr_val;
-        unsigned long long int num;
-        int i = 0, len = 0;
-        char hexnum[50];
+	unsigned long int temp_num = ptr_val;
+	unsigned long int num;
+	int i = 0, len = 0;
+	char hexnum[50];
 
-        while (temp_num != 0)
-        {
-                num = temp_num % 16;
-                if (num < 10)
-                        num += 48;
-                else
-                        num += 87;
-                hexnum[i] = num;
-                i++;
-                temp_num /= 16;
-        }
-        hexnum[i] = '\0';
-        while (hexnum[len] != '\0')
-                len++;
+	while (temp_num != 0)
+	{
+		num = temp_num % 16;
+		if (num < 10)
+			num += 48;
+		else
+			num += 87;
+		hexnum[i] = num;
+		i++;
+		temp_num /= 16;
+	}
+	hexnum[i] = '\0';
+	while (hexnum[len] != '\0')
+		len++;
 	_putchar('0');
 	_putchar('x');
 	count += 2;
-        for (i = len - 1; i >= 0; i--)
-        {
-                _putchar(hexnum[i]);
-                count++;
-        }
-        return (count);
+	for (i = len - 1; i >= 0; i--)
+	{
+		_putchar(hexnum[i]);
+		count++;
+	}
+	return (count);
 }
 /**
  * print_p - prints address of ptr in hex
@@ -45,9 +45,9 @@ int print_px(int count, unsigned long long int ptr_val)
 int print_p(int count, va_list list)
 {
 	void *temp = va_arg(list, void *);
-	unsigned long long int ptr_val;
+	unsigned long int ptr_val;
 
-	ptr_val = *(unsigned long long int *)&temp;
+	ptr_val = *(unsigned long int *)&temp;
 	count = print_px(count, ptr_val);
 	return (count);
 }
